+++ src/main/java/nl/lexemmens/podman/enumeration/PodmanCommand.java
@@ -1,24 +1,44 @@
 package nl.lexemmens.podman.enumeration;
 
-public enum PodmanCommand {
+public enum PodmanCommand implements Command {
 
     PODMAN("podman"),
     LOGIN("login"),
     BUILD("build"),
-    TAG("tag"),
-    SAVE("save"),
+    TAG("tag", false),
+    SAVE("save", false),
     PUSH("push"),
-    RMI("rmi"),
-    VERSION("version");
+    RMI("rmi", false),
+    VERSION("version", false, false);
 
-    private String command;
+    private final String command;
+    private final boolean tlsSupported;
+    private final boolean runRootSupported;
 
     PodmanCommand(String command) {
-        this.command = command;
+        this(command, true, true);
     }
 
+    PodmanCommand(String command, boolean tlsSupported) {
+        this(command, tlsSupported, true);
+    }
+
+    PodmanCommand(String command, boolean tlsSupported, boolean runRootSupported) {
+        this.command = command;
+        this.tlsSupported = tlsSupported;
+        this.runRootSupported = runRootSupported;
+    }
+
+    @Override
     public String getCommand() {
         return command;
     }
 
+    public boolean isTlsSupported() {
+        return tlsSupported;
+    }
+
+    public boolean isRunRootSupported() {
+        return runRootSupported;
+    }
 }