+++ src/main/java/nl/lexemmens/podman/service/BuildahExecutorService.java
@@ -3,11 +3,8 @@ package nl.lexemmens.podman.service;
 import nl.lexemmens.podman.enumeration.BuildahCommand;
 import nl.lexemmens.podman.executor.CommandExecutorDelegate;
 import nl.lexemmens.podman.config.podman.PodmanConfiguration;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.logging.Log;
-import org.zeroturnaround.exec.ProcessExecutor;
-import org.zeroturnaround.exec.stream.slf4j.Slf4jStream;
 
 import java.io.File;
 import java.util.ArrayList;
@@ -16,12 +13,8 @@ import java.util.List;
 /**
  * Class that allows very specific execution of Podman related commands.
  */
-public class BuildahExecutorService {
+public class BuildahExecutorService extends AbstractExecutorService<BuildahCommand> {
 
-    private static final File BASE_DIR = new File(".");
-
-    private final Log log;
-    private final CommandExecutorDelegate delegate;
     private final File podmanRoot;
 
 
@@ -33,8 +26,7 @@ public class BuildahExecutorService {
      * @param delegate     A delegate executor that executed the actual command
      */
     public BuildahExecutorService(Log log, PodmanConfiguration podmanConfig, CommandExecutorDelegate delegate) {
-        this.log = log;
-        this.delegate = delegate;
+        super(log, delegate);
         this.podmanRoot = podmanConfig.getRoot();
     }
 
@@ -58,36 +50,20 @@ public class BuildahExecutorService {
             return;
         }
 
-        List<String> subCommand = new ArrayList<>();
-        subCommand.add("rm");
-        subCommand.add("-rf");
-        subCommand.add(podmanRoot.getAbsolutePath());
-
-        runCommand(BuildahCommand.UNSHARE, subCommand);
+        command(BuildahCommand.UNSHARE)
+                .subCommand("rm")
+                .subCommand("-rf")
+                .subCommand(podmanRoot.getAbsolutePath())
+                .run();
     }
 
-    private List<String> decorateCommands(BuildahCommand podmanCommand, List<String> subCommands) {
+    @Override
+    protected List<String> compileCommandLine(BuildahCommand command, List<String> subCommands) {
         List<String> fullCommand = new ArrayList<>();
         fullCommand.add(BuildahCommand.BUILDAH.getCommand());
-        fullCommand.add(podmanCommand.getCommand());
+        fullCommand.add(command.getCommand());
         fullCommand.addAll(subCommands);
 
         return fullCommand;
-    }
-
-    private List<String> runCommand(BuildahCommand command, List<String> subCommands) throws MojoExecutionException {
-        List<String> fullCommand = decorateCommands(command, subCommands);
-
-        String msg = String.format("Executing command '%s' from basedir %s", StringUtils.join(fullCommand, " "), BASE_DIR.getAbsolutePath());
-        log.debug(msg);
-        ProcessExecutor processExecutor = new ProcessExecutor()
-                .directory(BASE_DIR)
-                .command(fullCommand)
-                .readOutput(true)
-                .redirectOutput(Slf4jStream.of(getClass().getSimpleName()).asInfo())
-                .redirectError(Slf4jStream.of(getClass().getSimpleName()).asError())
-                .exitValueNormal();
-
-        return delegate.executeCommand(processExecutor);
     }
 }